Feature/idcom 2186  cli implementation

+++ solana/program_v2/programs/gateway_v2/src/state/network.rs
@@ -69,11 +69,11 @@ impl GatekeeperNetwork {
             + OC_SIZE_U8 // auth_threshold
             + OC_SIZE_U64 // pass_expire_time
             + OC_SIZE_U8 // signer_bump
-            + OC_SIZE_VEC_PREFIX + NetworkFees::ON_CHAIN_SIZE * fees_count as usize // fees
-            + OC_SIZE_VEC_PREFIX + NetworkAuthKey::ON_CHAIN_SIZE * auth_keys as usize // auth_keys
-            + OC_SIZE_VEC_PREFIX + (OC_SIZE_PUBKEY * gatekeepers) as usize // gatekeeper list
+            + OC_SIZE_VEC_PREFIX + NetworkFees::ON_CHAIN_SIZE * fees_count // fees
+            + OC_SIZE_VEC_PREFIX + NetworkAuthKey::ON_CHAIN_SIZE * auth_keys // auth_keys
+            + OC_SIZE_VEC_PREFIX + (OC_SIZE_PUBKEY * gatekeepers) // gatekeeper list
             + OC_SIZE_U16 // network_index
-            + OC_SIZE_VEC_PREFIX + SupportedToken::ON_CHAIN_SIZE * supported_tokens as usize
+            + OC_SIZE_VEC_PREFIX + SupportedToken::ON_CHAIN_SIZE * supported_tokens
         // supported tokens list
     }
